Remove redundant MessagesContext (supposing to deal with messages in ChatsContext), and create src/tyoes folder to rename and move there UsersTypes, ChatsTypes, and MessagesTypes.
