+++ client/src/components/organisms/mains/Chat.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { useChatsContext } from "../../../contexts/chats/useChatsContext";
 import { useUsersContext } from "../../../contexts/users/useUsersContext";
 import MessageCard from "../../molecules/MessageCard";
-import { Chat as ChatType } from "../../../contexts/chats/chatsTypes";
+import { Chat as ChatT } from "../../../types/ChatsTypes";
 
 const StyledSection = styled.section`
   display: flex;
