+++ client/src/contexts/users/UsersContext.tsx
@@ -6,7 +6,7 @@ import {
   useCallback,
 } from "react";
 import { useNavigate } from "react-router-dom";
-import { UsersContextType, User, UsersState } from "./usersTypes";
+import { UsersContextType, User, UsersState } from "../../types/UsersTypes";
 import { usersReducer } from "./usersReducer";
 
 type UsersProviderProps = {
